add: improved goal action
